+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## V0.x-dev
+- Make configuration directory a constant
+- Change configuration directory from ~/.wp-sync-util to ~/.wp-site-sync
 
 ## V0.1.1
 - Added README.md file